Remove add registry, focus on endpoints
