+++ intellij-plugin/src/main/resources/inspectionDescriptions/ApolloEndpointNotConfigured.html
@@ -1,0 +1,13 @@
+<html>
+<body>
+Reports Apollo services that don't configure a GraphQL endpoint.
+<p>
+    Configuring a service's GraphQL endpoint URL is recommended, as it enables the execution of queries directly from the IDE.<br>
+    It also allows downloading the latest version of the schema easily, via a Gradle task or via the <kbd>Download Schema</kbd> action.<br>
+
+</p>
+<p>
+    <a href="https://www.apollographql.com/docs/kotlin/advanced/plugin-configuration#downloading-a-schema">More information</a>
+</p>
+</body>
+</html>
