--- intellij-plugin/src/main/resources/inspectionDescriptions/MissingIntrospection.html
@@ -1,11 +1,0 @@
-<html>
-<body>
-Reports Apollo services that don't configure an introspection or registry block.
-<p>
-    Configuring an introspection or registry block is recommended as it allows downloading the latest version of the schema easily via a Gradle task or via the <kbd>Download Schema</kbd> action.
-</p>
-<p>
-    <a href="https://www.apollographql.com/docs/kotlin/advanced/plugin-configuration#downloading-a-schema">More information</a>
-</p>
-</body>
-</html>
