+++ intellij-plugin/src/main/resources/messages/ApolloBundle.properties
@@ -121,8 +121,8 @@ inspection.unusedField.quickFix=Delete field
 inspection.apollo4Available.displayName=Apollo Kotlin 4 is available
 inspection.apollo4Available.reportText=Apollo Kotlin 4 is available
 inspection.apollo4Available.quickFix=Migrate to Apollo Kotlin 4
-inspection.missingIntrospection.displayName=GraphQL endpoint not configured
-inspection.missingIntrospection.reportText=GraphQL endpoint not configured
-inspection.missingIntrospection.quickFix.introspection=Add introspection block
+inspection.endpointNotConfigured.displayName=GraphQL endpoint not configured
+inspection.endpointNotConfigured.reportText=GraphQL endpoint not configured
+inspection.endpointNotConfigured.quickFix=Add introspection block
 
 notification.group.apollo=Apollo
