+++ intellij-plugin/src/test/testData/inspection/EndpointNotConfigured.gradle.kts
@@ -8,22 +8,6 @@ apollo {
     // Some comment
     codegenModels.set("operationBased")
     srcDir("src/main/graphql")
-    introspection {
-      endpointUrl.set("https://example.com/graphql")
-      schemaFile.set(file("src/main/graphql/schema.graphqls"))
-    }
-  }
-
-  service("b") { // Should highlight
-    packageName.set("com.example")
-    // Some comment
-    codegenModels.set("operationBased")
-    srcDir("src/main/graphql")
-    registry {
-      key.set(System.getenv("APOLLO_KEY"))
-      graph.set(System.getenv("APOLLO_GRAPH"))
-      schemaFile.set(file("src/main/graphql/schema.graphqls"))
-    }
   }
 
   service("c") {
@@ -37,7 +21,7 @@ apollo {
     }
   }
 
-  service("d") {
+  service("d") { // Should highlight
     packageName.set("com.example")
     // Some comment
     codegenModels.set("operationBased")
